refactor(pos): delete dead service-selector components, extract shared types
EnhancedServiceSelector, UnitItemSelector, WeightInput, ServiceTypeSelector,
and DynamicOrderItem were a fully-built alternate service-selection UI that
nothing renders - only their type exports were imported by the live POS flow
(EnhancedServiceSelectionPopup handles selection instead). Delete the dead
components and move the surviving types (Service, EnhancedOrderItem,
ServiceType, DynamicOrderItemData) into a shared orderTypes.ts, so
FloatingOrderSummary no longer redeclares an identical DynamicOrderItemData
interface locally.

Pure cleanup, no behavior change. First step of simplifying the order-
creation flow.
