+++ src/components/pos/EnhancedLaundryPOS.tsx
@@ -9,8 +9,7 @@ import { useCustomers } from '@/hooks/useCustomers';
 import { useCreateOrderWithNotifications as useCreateOrder, UnitItem } from '@/hooks/useOrdersWithNotifications';
 import { toast } from 'sonner';
 import { useServices, useSeedDefaultServices } from '@/hooks/useServices';
-import { EnhancedOrderItem } from './EnhancedServiceSelector';
-import { DynamicOrderItemData } from './DynamicOrderItem';
+import { EnhancedOrderItem, DynamicOrderItemData } from './orderTypes';
 import { EnhancedServiceSelectionPopup } from './EnhancedServiceSelectionPopup';
 import { FloatingOrderSummary } from './FloatingOrderSummary';
 import { CashPaymentDialog } from './CashPaymentDialog';
